add MVC and LabJackObject script
added the MVC and LabJackObject scripts, the LabJackObject is in the Overseer to try to get it connected across scenes. Code has been adapted from XchartLabJackTesting.cs to reproduce effects after cleaning.
Window sliding bug correction: serieMaxCache is now updated to match the selected visualisation window and not have data appear out of the chart